AHORA SI BACK FUNCIONANDO HASTA CARGA DE PEDIDOS

+++ BackEnd/src/main/java/com/buenSabor/BackEnd/models/user/Telefono.java
@@ -30,7 +30,7 @@ public class Telefono extends Bean {
 
 
     @Column(name = "numero")
-    private BigInteger numero;
+    private Long numero;
 
     @ManyToOne( fetch = FetchType.EAGER)
     @JsonIgnore
